interface/requirements.txt
Pillow
opencv-python
grpcio==1.39.0
grpcio-tools==1.39.0
tensorflow
pillow